regex que comprueba que el id es un numero y da un aviso si no lo es

+++ delete.php
@@ -1,5 +1,6 @@
 <?php
 include 'utils.php';
+include 'serv.php';
 session_start();
 if(!isset($_SESSION['usuario'])) {
     // no hay sesión iniciada
@@ -13,17 +14,26 @@ if(!isset($_SESSION['usuario'])) {
 }
 $_SESSION['tiempo'] = time();
 
-include 'serv.php';
-
-$sql = "DELETE FROM Jugadores WHERE ID = ?";
-# sentencia
-$sent = $conexion->prepare($sql);
-
-$sent->bind_param("i", $id);
 $id = $_GET['id'];
 
-$sent->execute();
-echo "<script> alert('Elemento borrado');</script>";
+# comprobar que el id es solo un numero y no una inyección sql
+$pattern = "/^[0-9]*$/";
+if (preg_match($pattern , $id)) {
+    echo "<script> alert('es un id lindo');</script>";
+    $sql = "DELETE FROM Jugadores WHERE ID = ?";
+    # sentencia
+    $sent = $conexion->prepare($sql);
+
+    $sent->bind_param("i", $id);
+    $id = $_GET['id'];
+
+    $sent->execute();
+    echo "<script> alert('Elemento borrado');</script>";
+} else {
+    $kanye_west_meme ="¯\\\_\(ツ\)_\/¯";
+    echo "<script> alert('Intento de inyección SQL detectado, su IP ha quedado registrada y se enviará a las autoridades.". $kanye_west_meme ."');</script>";
+}
+
 mysqli_close($conexion);
 echo '<script> window.location="panel.php"; </script>';
